requirements.txt
alembic==1.4.3
aniso8601==6.0.0
appdirs==1.4.4
asgiref==3.2.10
atomicwrites==1.4.0
attrs==20.3.0
awscli==1.18.190
Babel==2.8.0
botocore==1.19.30
certifi==2020.11.8
cffi==1.14.3
chardet==3.0.4
Click==7.0
clock==0.1
colorama==0.4.3
colorclass==2.2.0
cryptography==3.1.1
distlib==0.3.1
Django==3.1
docopt==0.6.2
docutils==0.15.2
ecdsa==0.14.1
filelock==3.0.12
Flask==1.0.3
flask-core==2.9.0
Flask-Cors==3.0.7
Flask-Migrate==2.5.3
Flask-Moment==0.10.0
Flask-RESTful==0.3.7
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.0
Flask-WTF==0.14.3
gunicorn==20.0.4
idna==2.10
iniconfig==1.1.1
itsdangerous==1.1.0
Jinja2==2.10.1
jmespath==0.10.0
lazy-object-proxy==1.4.0
Mako==1.1.3
MarkupSafe==1.1.1
packaging==20.4
pip-upgrader==1.4.15
pluggy==0.13.1
postgres==3.0.0
psycopg2==2.8.6
psycopg2-binary==2.8.6
psycopg2-pool==1.1
py==1.9.0
pyasn1==0.4.8
pycparser==2.20
pycryptodome==3.9.9
PyJWT==1.7.1
pyparsing==2.4.7
pytest==6.1.2
python-dateutil==2.6.0
python-editor==1.0.4
python-jose==3.2.0
pytz==2019.1
PyYAML==5.3.1
requests==2.25.0
rsa==4.5
s3transfer==0.3.3
six==1.12.0
SQLAlchemy==1.3.4
sqlparse==0.3.1
terminaltables==3.1.0
toml==0.10.2
tzlocal==2.1
urllib3==1.26.2
virtualenv==20.0.25
Werkzeug==0.15.5
wrapt==1.11.1
WTForms==2.3.3
